Add files via upload
Here, we use basic principles of recursion to create a Sierpinski Triangle, which consists of an equilateral triangle of unit size, subdivided into 4 smaller equilateral triangles. We first draw a triangle of half the size inside the first triangle with its apex at the midpoint of its base, and its base joining the midpoints of its other two sides. This central triangle splits the remaining triangle into three others, each with one of the original triangle's vertices as their apices. We repeat the above procedure recursively for each of them to our desired depth.